requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
itypes==1.2.0
jinja2==3.1.4
markupsafe==2.1.5
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
